% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population data}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	4060 obs. of  3 variables:
 $ country   : chr  "Afghanistan" "Afghanistan" "Afghanistan" "Afghanistan" ...
 $ year      : int  1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 ...
 $ population: int  17586073 18415307 19021226 19496836 19987071 20595360 21347782 22202806 23116142 24018682 ...
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
population
}
\description{
Populations of 219 countries for 1995-2013.
}
\keyword{datasets}

