% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table1}
\alias{table1}
\alias{table2}
\alias{table3}
\alias{table4}
\alias{table5}
\title{tabular layouts}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	9 obs. of  4 variables:
 $ country   : chr  "Afghanistan" "Afghanistan" "Afghanistan" "Brazil" ...
 $ year      : int  1999 2000 2001 1999 2000 2001 1999 2000 2001
 $ cases     : int  745 2666 4639 37737 80488 37491 212258 213766 212766
 $ population: int  19987071 20595360 21347782 172006362 174504898 176968205 1272915272 1280428583 1287890449
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
table1
}
\description{
Data sets that demonstrate multiple ways to layout the same tabular data.
}
\details{
\code{table1}, \code{table2}, \code{table3}, \code{table4}, and \code{table5} all display the number of TB cases documented by the World Health Organization in Afghanistan, Brazil, and China between 1999 and 2001. The data contains values associated with four variables (country, year, cases, and population), but each table organizes the values in a different layout.

The data is a subset of the data contained in the World Health
Organization Global Tuberculosis Report
}
\keyword{datasets}

